(CTH-174) Add Connector
Adding connector/connector and connector/timing (old source files
from src/websocket of cthun-agent).

The logic for the login is now included in the onOpen handler. We have a
new postLogin callback that is executed after the login message is sent.

As a consequence of including the login logic, the Connector class now
includes message.h and data_container.h; it's not a WebSocket wrapper
anymore - it knows about Cthun communications.

Adding new CloseCode type; we're exposing only three close codes:
normal, abnormal_close, subprotocol_error.

Adding the monitorConnectionState method; it was previously a method of
cthun-agent's Agent class.