(CTH-191) Add Schema parsing constructor
Adding a ctor that parses a json schema. Note that, due to validjson
limitation, you can't add constraints after using that ctor.

Making DataContainer::getRaw a const method.

When using valijson populateSchema() to create a schema, it is not
possible to add further constraints; we explicitly forbid that by
throwing a new schema_error.