(maint) Add unit tests for Connection and ClientMetadata
Also fixing ClientMetadata; previously, it was setting "agent" as
the type suffix in the id string - it now uses the provided type.

Improving error handling of ClientMetadata; the openSSL-based function
that retrieves the client identity from the certificate file was
previously giving a seg fault in case the specified file did not exist.