(CTH-193) Report parsing errors in ParsedChunks
Message::getParsedChunks() does not propagte parsing errors for data and
debug chunks anymore; they are now stored in ParseChunks objects.

This allows the RPC callbacks to process such ParsedChunks objects and
reply to the requester client. Previously, the RPC callbacks were not
triggered; the Connector::processMessage callback simply returned after
a parsing failure.

The changes consist in modifying the ParsedChunks struct and the
Message::getParsedChunks() logic.