(PCP-53) Port cpp-pcp-client to boost::thread
Here we add two header files to encapsulate which threading, locking and
chrono libraries we use. Due to us not currently having working
std::thread available on AIX, we make the switch to boost::thread. In
the future we can switch back to using the standard libraries by
replacing boost:: with std:: in thread.hpp and chrono.hpp.