(PCP-177) Allow API user to set connection timeout
websocketpp uses a 5000ms default when setting the timeout for
establishing a connection. While it isn't clear from the code and
comments in the library, this timeout includes dns, connecting, tls,
and the websocket handshake.

Here we make that timeout configurable buy adding an extra parameter to
the Connector ctor which we use in the connection object to set
set_open_handshake_timeout. We still default to 5000ms.