(PCP-216) Connection FSM waits for onOpen after connect
With this commit we improve the way the Finite State Machine of
Connection::connect() deals with onOpen events by taking into account the
connection timeout configured for the underlaying websocketpp connection
instance and checking the current state of the connection.

Also, in Connection::cleanup(), we now consider the current state of the
connection before calling stop_perpetual() for the underlaying
websocketpp endpoint instance; in case it's on 'connecting' state, the
method will wait for the connection timeout interval mentioned above.

The above changes prevent the underlaying WebSocket event loop thread
to continue after Connection::connect() throws a connection_fatal_error,
due to concurrent WebSocket events being handled while the Connection
dtor executes. On the other hand, in case of connection failures that
don't trigger an onFail event when the max num of connection attempts
is set, the Connection dtor won't return immediately.