(PCP-371) Spike i18n support
Enable the creation of the .pot file that includes the strings that
we want to translate. Don't enable localization now; we'll do that on
PCP-257.

Use leatherman::locale::format() and translate() for the strings that we
want to translate.

Use "{<int>}" instead of "%<int>%" for string interpolation.

Don't use leatherman::util::plural(); we'll deal with plurals once we
enable localization on PCP-257.

Moving ConnectionTimings::toString() to timings.cc, in order to enable
i18n.