(PCP-383) Implement broker failover
Implements sticky broker failover when given a list of brokers. It will
attempt to connect to the first broker, if it fails try the next broker,
if that fails try the next, and so on, cycling through the list.

If a prior connection existed and failed, it will retry that connection,
then if that fails cycle through the list starting after that entry.