(maint) Default to pass constructor arguments by value
When passing arguments to a constructor that will be stored directly
into the object, pass by value. This allows move semantics while not
incurring any overhead if we need to create a copy (the copy will be
created on constructor invocation rather than in the constructor). This
requires that the argument is then `std::move`d into the object storage.

In most other cases, preserve pass-by-reference.

See
https://github.com/puppetlabs/cppbestpractices/blob/master/03-FAQ.md#q-arguments-and-return-types-in-class-methods