(PCP-383) Add simple test for failover against mock server
Create a websocket server for testing against. Uses test certs with
CN=localhost so we can test successful TLS connections. The mock server
dynamically assigns a port, so we can be sure we get a socket.

Adds a simple test for failing over to a backup when the primary broker
is down.

Dynamically assigning ports required updating to websocketpp 0.7.0 to
pick up the `get_local_endpoint` method.