(PCP-425) Wait before closing connection if connecting
Setting the connection state to closing in Connection::close(); using a
mutex to handle the connection state.

Improving the logic for checking the connection state when the
Connection::clean() is invoked (ex. by the dtor): wait in case the
connection state is connecting or closing. Note that the underlying
websocketpp won't close the connection if the WebSocket onOpen event has
not being triggered.