(PCP-455) Add unit tests for Connector
Adding new functionalities for the MockServer:
 - configurable handler for WebSocket pings;
 - Association Request handler (triggered by onMessage event).

The connection_check interval of Connector::startMonitoring is now an
argument of the method; this allows us to write unit tests with a
monitor task that has a very short period.

Adding new connector_test.cc with unit tests for Connector::connect and
Connector::startMonitoring (to ensure it reconnects when the broker goes
down and that the connection gets closed after a number of pong timeouts).

Configuring the WebSocket pong timeout via Connector ctor.

Adding new PRIVATE pcp-chunks library, to provide the nedded
cpp-pcp-client functionalities to the mock-server library.