(PCP-411) Retrieve and expose PCP Association timings
Adding new AssociationTimings struct, to keep track of the time spent
for associating and the entire session duration.

Adding AssociationTimings member to Connector.

Adding methods to set onClose and onFail callbacks to Connection.

Resetting and updating AssociationTimings in Connector's onOpen callback
(the one that sends that Associate request) and in the connect() synch
method.

Adding AssociationTimings getter to Connector.

Adding AssociationTimings unit tests to connector_test.