(PCP-738) increase randomization of PCP connection retry
Previously the retry logic would always retry connections within 250ms
of 2, 4, 8, etc. This is insufficient to prevent a thundering herd when
large numbers of agents are involved. This change carries randomization
from the nth iteration into the n+1th, which should increase the
divergence between agents as number of retries increases.