(maint) Move setting CC/CXX to .travis.yml for coverity
Moves setting the CC/CXX environment variables to .travis.yml to avoid
merge conflicts when merging to coverity_scan. Placing these variables
in the same file as we install the packages makes it clearer how they're
being configured, and ensures they're set for build steps that don't use
the script.