(maint) Make static lib the default, remove EL4 hack
EL4 required pthreads be used only in a shared library. Since we no
longer support EL4, remove the hack to produce a shared library for
mock-server and put all test code in a single binary. We can now use a
single static or shared library for libcpp-pcp-client that we install
and test.

Also switch to producing a static library by default as it's most
likely to be used in a single binary.