(packaging) Use packaging as a gem
This commit removes the packaging bootstrap and implode rake tasks and replaces
them with the packaging gem. Since this project doesn't have a top-level
Gemfile, we must install the packaging gem ourselves (in a job, locally, etc.)
in order to make use of the packaging rake tasks (which should only be needed
at ship time for this project).