(PCP-906) Thread log level/stream through connector constructor
These bits of data need to be stored on the connector so they can be
configured on the connection once it's been created. We need to do
that as soon as the connection is created and before it connects (so we
get logs associated with the connect process), which means it has to
be handled by the connector itself, and not called from the outside
after the fact.