(PE-33165) Disable websocket logging
There is an undiagnosed bug around websocket logging on Windows, so
temporarily disable that logging until a solution can be found.

This restores the use of the default ASIO TLS client config, rather than
our custom config which enables devel logging. It also explicitly sets
the websocket log level to none and doesn't set a stream.